ngspice_cloud/requirements.txt
asgiref==3.2.3
cffi==1.13.2
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
kiwisolver==1.1.0
matplotlib==3.1.2
ngspicepy==1.0
numpy==1.18.1
ply==3.11
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
PySpice==1.3.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
